[backend] Extract Kanban service and Pipeline services
