+++ backend/src/main/java/heartbeat/config/WebConfig.java
@@ -2,9 +2,8 @@ package heartbeat.config;
 
 import heartbeat.controller.board.dto.request.BoardType;
 import heartbeat.controller.pipeline.dto.request.PipelineType;
-import heartbeat.controller.report.dto.request.DataType;
-import heartbeat.controller.report.dto.request.MetricType;
 import heartbeat.controller.report.dto.request.ReportType;
+import heartbeat.controller.report.dto.request.ReportDataType;
 import heartbeat.controller.source.SourceType;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.convert.converter.Converter;
@@ -30,17 +29,17 @@ public class WebConfig implements WebMvcConfigurer {
 			}
 		});
 
-		registry.addConverter(new Converter<String, MetricType>() {
-			@Override
-			public MetricType convert(String type) {
-				return MetricType.fromValue(type);
-			}
-		});
-
 		registry.addConverter(new Converter<String, ReportType>() {
 			@Override
 			public ReportType convert(String type) {
 				return ReportType.fromValue(type);
+			}
+		});
+
+		registry.addConverter(new Converter<String, ReportDataType>() {
+			@Override
+			public ReportDataType convert(String type) {
+				return ReportDataType.fromValue(type);
 			}
 		});
