--- backend/src/main/java/heartbeat/controller/report/dto/request/ExportCSVRequest.java
@@ -1,18 +1,0 @@
-package heartbeat.controller.report.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class ExportCSVRequest {
-
-	private String dataType;
-
-	private String csvTimeStamp;
-
-}
