+++ backend/src/main/java/heartbeat/controller/report/dto/request/ReportType.java
@@ -2,23 +2,12 @@ package heartbeat.controller.report.dto.request;
 
 public enum ReportType {
 
-	METRIC("metric"), PIPELINE("pipeline"), BOARD("board");
-
-	private String value;
-
-	ReportType(String value) {
-		this.value = value;
-	}
-
-	public String getValue() {
-		return value;
-	}
+	BOARD, DORA;
 
 	public static ReportType fromValue(String type) {
 		return switch (type) {
-			case "metric" -> METRIC;
-			case "pipeline" -> PIPELINE;
 			case "board" -> BOARD;
+			case "dora" -> DORA;
 			default -> throw new IllegalArgumentException("ReportType not found!");
 		};
 	}
