+++ backend/src/main/java/heartbeat/controller/report/dto/response/ReportResponse.java
@@ -6,6 +6,9 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.util.List;
+import java.util.function.Function;
+
+import static heartbeat.util.ValueUtil.getValueOrNull;
 
 @Data
 @Builder
@@ -41,6 +44,26 @@ public class ReportResponse {
 
 	public ReportResponse(Long exportValidityTime) {
 		this.exportValidityTime = exportValidityTime;
+	}
+
+	public ReportResponse(ReportResponse boardReportResponse, ReportResponse doraReportResponse,
+			ReportResponse codebaseReportResponse, MetricsDataCompleted metricsDataCompleted, ReportResponse response,
+			ReportError reportError, boolean isReportReady) {
+		this.velocity = getValueOrNull(boardReportResponse, ReportResponse::getVelocity);
+		this.classificationList = getValueOrNull(boardReportResponse, ReportResponse::getClassificationList);
+		this.cycleTime = getValueOrNull(boardReportResponse, ReportResponse::getCycleTime);
+		this.exportValidityTime = getValueOrNull(response, ReportResponse::getExportValidityTime);
+		this.deploymentFrequency = getValueOrNull(doraReportResponse, ReportResponse::getDeploymentFrequency);
+		this.changeFailureRate = getValueOrNull(doraReportResponse, ReportResponse::getChangeFailureRate);
+		this.meanTimeToRecovery = getValueOrNull(doraReportResponse, ReportResponse::getMeanTimeToRecovery);
+		this.leadTimeForChanges = getValueOrNull(codebaseReportResponse, ReportResponse::getLeadTimeForChanges);
+		this.boardMetricsCompleted = getValueOrNull(metricsDataCompleted, MetricsDataCompleted::boardMetricsCompleted);
+		this.pipelineMetricsCompleted = getValueOrNull(metricsDataCompleted,
+				MetricsDataCompleted::pipelineMetricsCompleted);
+		this.sourceControlMetricsCompleted = getValueOrNull(metricsDataCompleted,
+				MetricsDataCompleted::sourceControlMetricsCompleted);
+		this.allMetricsCompleted = isReportReady;
+		this.reportError = reportError;
 	}
 
 }
