+++ backend/src/main/java/heartbeat/service/report/CSVFileGenerator.java
@@ -4,7 +4,7 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.opencsv.CSVWriter;
 import heartbeat.controller.board.dto.response.JiraCardDTO;
-import heartbeat.controller.report.dto.request.ReportType;
+import heartbeat.controller.report.dto.request.ReportDataType;
 import heartbeat.controller.report.dto.response.BoardCSVConfig;
 import heartbeat.controller.report.dto.response.LeadTimeInfo;
 import heartbeat.controller.report.dto.response.PipelineCSVInfo;
@@ -36,7 +36,6 @@ import org.apache.commons.lang3.StringUtils;
 import org.springframework.core.io.InputStreamResource;
 import org.springframework.stereotype.Component;
 
-import java.io.ByteArrayInputStream;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileWriter;
@@ -138,8 +137,8 @@ public class CSVFileGenerator {
 		}
 	}
 
-	public InputStreamResource getDataFromCSV(ReportType reportType, long csvTimeStamp) {
-		return switch (reportType) {
+	public InputStreamResource getDataFromCSV(ReportDataType reportDataType, long csvTimeStamp) {
+		return switch (reportDataType) {
 			case METRIC -> readStringFromCsvFile(
 					CSVFileNameEnum.METRIC.getValue() + FILENAME_SEPARATOR + csvTimeStamp + CSV_EXTENSION);
 			case PIPELINE -> readStringFromCsvFile(
