+++ backend/src/main/java/heartbeat/util/ValueUtil.java
@@ -1,0 +1,15 @@
+package heartbeat.util;
+
+import java.util.function.Function;
+
+public class ValueUtil {
+
+	public static <T, R> R getValueOrNull(T object, Function<T, R> getter) {
+		return object != null ? getter.apply(object) : null;
+	}
+
+	public static <T> T valueOrDefault(T defaultValue, T value) {
+		return value == null ? defaultValue : value;
+	}
+
+}
