+++ backend/src/test/java/heartbeat/controller/report/dto/request/ReportTypeTest.java
@@ -3,19 +3,17 @@ package heartbeat.controller.report.dto.request;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class ReportTypeTest {
 
 	@Test
 	public void shouldConvertValueToType() {
 		ReportType boardType = ReportType.fromValue("board");
-		ReportType pipelineType = ReportType.fromValue("pipeline");
-		ReportType metricType = ReportType.fromValue("metric");
+		ReportType doraType = ReportType.fromValue("dora");
 
 		assertEquals(boardType, ReportType.BOARD);
-		assertEquals(pipelineType, ReportType.PIPELINE);
-		assertEquals(metricType, ReportType.METRIC);
+		assertEquals(doraType, ReportType.DORA);
 	}
 
 	@Test
