+++ backend/src/test/java/heartbeat/service/report/CalculateLeadTimeForChangesTest.java
@@ -8,6 +8,8 @@ import heartbeat.controller.report.dto.response.AvgLeadTimeForChanges;
 import heartbeat.controller.report.dto.response.LeadTimeForChanges;
 import heartbeat.controller.report.dto.response.LeadTimeForChangesOfPipelines;
 import java.util.List;
+
+import heartbeat.service.report.calculator.LeadTimeForChangesCalculator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
